+++ Engine/Private/ParticleManager.cpp
@@ -158,7 +158,7 @@ HRESULT CParticleManager::Render(ID3D11DeviceContext* pContext, const RENDER_CTX
                 TAG_RES_GRP_PERMANENT_STATE, TAG_RES_STATE_RS_SOLID_BACKCULL);
             pContext->RSSetState(rasterizer->GetRasterizerState().Get());
         }
-        pContext->Draw(m_arrVertices[ETOUI(PARTICLE_TYPE::BLOCK_DESTRUCT)].size(), 0);
+        pContext->Draw((uint32_t)m_arrVertices[ETOUI(PARTICLE_TYPE::BLOCK_DESTRUCT)].size(), 0);
 
         pContext->GSSetShader(nullptr, nullptr, 0);
